+++ client/app/pricing/page.tsx
@@ -1,0 +1,5 @@
+import Pricing from "../../src/views/Pricing";
+
+export default function PricingPage() {
+  return <Pricing />;
+}